requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas
numpy
matplotlib
scikit-learn
joblib

# dacot>=2.0.0

# Mapping
plotly
dash
